requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
cryptography==37.0.2
EasyProcess==1.1
h11==0.13.0
idna==3.3
MouseInfo==0.1.3
outcome==1.1.0
Pillow==9.1.1
PyAutoGUI==0.9.53
pycparser==2.21
pyenchant==3.2.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python3-xlib==0.15
pytweening==1.0.4
selenium==4.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
pexpect~=4.8.0
psutil~=5.9.1
PyPDF2~=1.28.4
mysql-connector==2.2.9